pyproject.toml
[tool.poetry]
name = "aininjas"
version = "0.1.0"
description = ""
authors = ["bryan Ninja AI <bryanNINJA@ikigaidojo.ninja>"]
readme = "README.md"

[tool.poetry.dependencies]
advertools = "^0.13.2"
argon2-cffi = "21.3.0"
asgiref = "^3.6.0"
boto3 = "^1.26.80"
celery = "5.2.7"
crispy-bootstrap5 = "0.7"
django = "4.1.5"
django-allauth = "0.52.0"
django-anymail = {extras = ["production"], version = "9.0"}
django-celery-beat = "2.4.0"
django-cors-headers = "3.13.0"
django-crispy-forms = "1.14.0"
django-environ = "0.9.0"
django-extensions = "3.2.1"
django-health-check = "^3.17.0"
django-model-utils = "4.3.1"
django-ninja = "0.20.0"
django-ninja-extra = "0.18.4"
django-redis = "5.2.0"
djfernet = "^0.8.1"
emoji = "2.2.0"
flower = "1.2.0"
geckodriver-autoinstaller = "0.1.0"
google-api-python-client = "^2.82.0"
google-auth = "^2.16.3"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
gspread = "5.7.2"
gunicorn = {version = "20.1.0", extras = ["production"]}
hiredis = "2.1.0"
langchain = "^0.0.105"
merry = "0.3.0"
nltk = "3.8.1"
oauth2client = "4.1.3"
openai = "0.27.0"
openai-async = "^0.0.3"
pandas = "1.5.2"
pillow = "9.4.0"
plantuml = "0.3.0"
psycopg2 = "2.9.5"
python = "3.10.5"
python-slugify = "7.0.0"
python-youtube = "0.9.0"
pytube = "12.1.2"
pytz = "2022.7"
redis = "4.4.1"
revchatgpt = "2.3.5"
scrapetube = "2.3.1"
sentry-sdk = {version = "1.12.1", extras = ["production"]}
snscrape = "0.5.0.20230113"
spacy = "3.5.0"
supabase = "1.0.2"
tiktoken = "^0.3.1"
tweepy = "4.12.1"
whitenoise = "6.3.0"
youtube-transcript-api = "0.5.0"


[tool.poetry.group.dev.dependencies]
selenium = "4.7.2"
werkzeug = {extras = ["watchdog"], version = "2.2.2"}
ipdb = "0.13.11"
watchfiles = "0.18.1"
mypy = "0.991"
django-stubs = "1.13.1"
pytest = "7.2.0"
pytest-sugar = "0.9.6"
sphinx = "6.1.2"
sphinx-autobuild = "2021.3.14"
flake8 = "6.0.0"
flake8-isort = "6.0.0"
coverage = "7.0.4"
black = "22.12.0"
pylint-django = "2.5.3"
pylint-celery = "0.3"
pre-commit = "2.21.0"
factory-boy = "3.2.1"
django-debug-toolbar = "3.8.1"
django-extensions = "3.2.1"
django-coverage-plugin = "3.0.0"
pytest-django = "4.5.2"
debugpy = "1.6.5"
snoop = "0.4.3"
pysnooper = "1.1.1"
pytest-mock = "3.10.0"
pytest-watch = "4.2.0"
pytest-asyncio = "0.20.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
